backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.32.0
python-multipart==0.0.18
pydantic==2.10.4
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
requests==2.31.0
